quant_analysis/requirements.txt
yfinance
pandas
numpy
yahoo_fin
requests
sentence-transformers
faiss-cpu
groq
openai

